Extract variable context guard to separate module
Makes the variable module more structured.
