+++ yash-env/src/variable.rs
@@ -44,6 +44,7 @@
 //! [`Env::push_context`] returns a [`EnvContextGuard`] that implements
 //! `DerefMut<Target = Env>`.
 
+#[cfg(doc)]
 use crate::Env;
 use either::{Left, Right};
 use itertools::Itertools;
@@ -54,8 +55,6 @@ use std::ffi::CString;
 use std::fmt::Write;
 use std::hash::Hash;
 use std::iter::FusedIterator;
-use std::ops::Deref;
-use std::ops::DerefMut;
 use thiserror::Error;
 use yash_syntax::source::Location;
 
@@ -847,107 +846,9 @@ impl<'a> Iterator for Iter<'a> {
 
 impl FusedIterator for Iter<'_> {}
 
-/// RAII-style guard for temporarily retaining a variable context.
-///
-/// The guard object is created by [`VariableSet::push_context`].
-#[derive(Debug)]
-#[must_use = "You must retain ContextGuard to keep the context alive"]
-pub struct ContextGuard<'a> {
-    stack: &'a mut VariableSet,
-}
+mod guard;
 
-impl VariableSet {
-    /// Pushes a new empty context to this variable set.
-    ///
-    /// This function returns a scope guard that will pop the context when dropped.
-    #[inline]
-    pub fn push_context(&mut self, context_type: ContextType) -> ContextGuard<'_> {
-        self.push_context_impl(context_type);
-        ContextGuard { stack: self }
-    }
-
-    /// Pops the topmost context from the variable set.
-    #[inline]
-    pub fn pop_context(guard: ContextGuard<'_>) {
-        drop(guard)
-    }
-}
-
-/// When the guard is dropped, the context that was pushed when creating the
-/// guard is [popped](VariableSet::pop_context).
-impl std::ops::Drop for ContextGuard<'_> {
-    #[inline]
-    fn drop(&mut self) {
-        self.stack.pop_context_impl()
-    }
-}
-
-impl std::ops::Deref for ContextGuard<'_> {
-    type Target = VariableSet;
-    #[inline]
-    fn deref(&self) -> &VariableSet {
-        self.stack
-    }
-}
-
-impl std::ops::DerefMut for ContextGuard<'_> {
-    #[inline]
-    fn deref_mut(&mut self) -> &mut VariableSet {
-        self.stack
-    }
-}
-
-/// RAII-style guard that makes sure a context is popped properly
-///
-/// The guard object is created by [`Env::push_context`].
-#[derive(Debug)]
-#[must_use = "The context is popped when the guard is dropped"]
-pub struct EnvContextGuard<'a> {
-    env: &'a mut Env,
-}
-
-impl Env {
-    /// Pushes a new context to the variable set.
-    ///
-    /// This function is equivalent to
-    /// `self.variables.push_context(context_type)`, but returns an
-    /// `EnvContextGuard` that allows re-borrowing the `Env`.
-    #[inline]
-    pub fn push_context(&mut self, context_type: ContextType) -> EnvContextGuard<'_> {
-        self.variables.push_context_impl(context_type);
-        EnvContextGuard { env: self }
-    }
-
-    /// Pops the topmost context from the variable set.
-    #[inline]
-    pub fn pop_context(guard: EnvContextGuard<'_>) {
-        drop(guard)
-    }
-}
-
-/// When the guard is dropped, the context that was pushed when creating the
-/// guard is [popped](VariableSet::pop_context).
-impl Drop for EnvContextGuard<'_> {
-    #[inline]
-    fn drop(&mut self) {
-        self.env.variables.pop_context_impl()
-    }
-}
-
-impl Deref for EnvContextGuard<'_> {
-    type Target = Env;
-    #[inline]
-    fn deref(&self) -> &Env {
-        self.env
-    }
-}
-
-impl DerefMut for EnvContextGuard<'_> {
-    #[inline]
-    fn deref_mut(&mut self) -> &mut Env {
-        self.env
-    }
-}
+pub use self::guard::{ContextGuard, EnvContextGuard};
 
 #[cfg(test)]
 mod tests {
@@ -1625,41 +1526,5 @@ mod tests {
         assert_matches!(&variables.positional_params().value, Some(Array(values)) => {
             assert_eq!(values.as_ref(), ["x".to_string()]);
         });
-    }
-
-    #[test]
-    fn scope_guard() {
-        let mut env = Env::new_virtual();
-        let mut guard = env.variables.push_context(ContextType::Regular);
-        guard
-            .assign(Scope::Global, "foo".to_string(), Variable::new(""))
-            .unwrap();
-        guard
-            .assign(Scope::Local, "bar".to_string(), Variable::new(""))
-            .unwrap();
-        VariableSet::pop_context(guard);
-
-        let variable = env.variables.get("foo").unwrap();
-        assert_eq!(variable.value, Some(Value::scalar("")));
-        assert_eq!(env.variables.get("bar"), None);
-    }
-
-    #[test]
-    fn env_scope_guard() {
-        let mut env = Env::new_virtual();
-        let mut guard = env.push_context(ContextType::Regular);
-        guard
-            .variables
-            .assign(Scope::Global, "foo".to_string(), Variable::new(""))
-            .unwrap();
-        guard
-            .variables
-            .assign(Scope::Local, "bar".to_string(), Variable::new(""))
-            .unwrap();
-        Env::pop_context(guard);
-
-        let variable = env.variables.get("foo").unwrap();
-        assert_eq!(variable.value, Some(Value::scalar("")));
-        assert_eq!(env.variables.get("bar"), None);
     }
 }
